requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
assemblyai==0.25.0
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.2
caer==2.0.8
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
contourpy==1.1.0
cvzone==1.6.1
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
elevenlabs==0.3.0b0
executing==1.2.0
fastapi==0.104.0
fastjsonschema==2.18.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
lab==7.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
libclang==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.0.1
MouseInfo==0.1.3
multidict==6.0.5
mypy==1.7.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
openai==1.12.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
panda==0.3.1
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pypdfium2==4.27.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
setuptools-rust==1.8.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.1
SQLAlchemy==2.0.27
stack-data==0.6.2
starlette==0.27.0
streamlit==1.31.1
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
txt2tags==3.8
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
vaderSentiment==3.3.2
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==12.0
widgetsnbextension==4.0.8
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0
